Fix linting errors
