+++ src/Components/LoginView.jsx
@@ -5,12 +5,11 @@ import { Button, Image } from 'semantic-ui-react';
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 import logo from '../images/slack-logo.png';
-//import slackConnectHref from '../Constants/connectWithSlackHref';
+// import slackConnectHref from '../Constants/connectWithSlackHref';
 import slack from '../images/slackIcon.png';
-import { connectWithSlack } from '../Actions/index.js';
+import { connectWithSlack } from '../Actions/index';
 
 export function LoginView(props) {
-  // console.log(props)
   const { connectWithSlack } = props;
   return (
     <div>
