+++ src/Components/Message.jsx
@@ -2,8 +2,8 @@
 
 import React from 'react';
 import { Image, List } from 'semantic-ui-react';
+import type { MessageType } from '../FlowTypes/Types';
 
-import type { MessageType } from './store';
 
 export default function Message(
   { avatarImage, name, text, timestamp }: MessageType,
